feat: update the name from actions to api
