+++ packages/admin/src/app/(dashboard)/playground/event/[...event]/layout.tsx
@@ -3,10 +3,10 @@ import { ReactNode } from 'react';
 
 import { framework } from '@/lib/framework-utils';
 
-import { EventPlaygroundProvider } from '@/domains/playground/providers/event-playground-provider';
+import { EventPlaygroundProvider } from '@/domains/playground/context/event-playground-context';
 import { getSerializedFrameworkEvents } from '@/domains/workflows/utils';
 
-import { PlaygroundBreadCrumb } from '../../components/playground-breadcrumb';
+import { PlaygroundBreadCrumb } from '@/domains/playground/components/playground-breadcrumb';
 
 export default async function Layout({ children }: { children: ReactNode }) {
   const systemEvents = framework?.getSystemEvents();
